Files renamed according to #1 schema
